Fix "always Auto Switch" was disabled in dialog where it was enabled
IniWrite was extracted in commit dd751a0337335b0436bb28d9a9918cf5a6981ef8 to reduce disk usage. The default DialogAction value was written to the INI for the first time (it wasnt written before this commit). Then it was permanently read for this dialog. Since the parameter existed in the INI, it was considered overriding for this dialog and the AutoSwitch value was ignored.