Add multi-dimensional arrays
Array for AddMenuPaths() must be array of arrays: [path, icon, iconNumber?, title?]

If the `title` of menu item is specified, it will be used instead of short path.

`Icon` is abosolute / relative to IconsDir path to `ICO, CUR, ANI, EXE, DLL, CPL, SCR` and other resource that contains icons

To use an icon group other than the first one in the file, specify its number for IconNumber *(if omitted, it defaults to 1)*. For example, 2 would load the default icon from the second icon group. If `IconNumber` is negative, its absolute value is assumed to be the resource ID of an icon within an executable file.